% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{autoinst}
\alias{autoinst}
\title{Automatically install missing packages when they are not found}
\usage{
autoinst()
}
\description{
This function is intended to be used as an error option. It can be set using
\code{options(error = autoinst)}.
}
